ai-apps/langsmith-inv-app/requirements.txt
langchain==1.2.6
langchain-anthropic==1.3.1
langchain-google-genai==4.2.0
langchain-openai==1.1.7
langsmith==0.6.4
python-dotenv==1.2.1
streamlit==1.53.0
